requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
amqp==5.1.1
    # via kombu
asgiref==3.7.2
    # via django
async-timeout==4.0.3
    # via redis
billiard==4.1.0
    # via celery
celery==5.3.4
    # via -r requirements.in
click==8.1.7
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.0
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.3.0
    # via celery
django==4.2.6
    # via
    #   -r requirements.in
    #   djangorestframework
djangorestframework==3.14.0
    # via -r requirements.in
kombu==5.3.2
    # via celery
prompt-toolkit==3.0.39
    # via click-repl
python-dateutil==2.8.2
    # via celery
pytz==2023.3.post1
    # via djangorestframework
redis==5.0.1
    # via -r requirements.in
six==1.16.0
    # via python-dateutil
sqlparse==0.4.4
    # via django
typing-extensions==4.8.0
    # via asgiref
tzdata==2023.3
    # via celery
vine==5.0.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.8
    # via prompt-toolkit
